+++ tomcat/src/main/java/org/apache/coyote/http11/handler/HandlerMapper.java
@@ -1,9 +1,9 @@
 package org.apache.coyote.http11.handler;
 
-import static org.apache.coyote.http11.handler.HandlerForGetRequest.HANDLER_FOR_GET_REQUEST;
 import static org.apache.coyote.http11.handler.LoginHandler.LOGIN_HANDLER;
-import static org.apache.coyote.http11.handler.LoginPageHandler.LOGIN_PAGE_HANDLER;
 import static org.apache.coyote.http11.handler.RegisterHandler.REGISTER_HANDLER;
+import static org.apache.coyote.http11.handler.get.HandlerForGetRequest.HANDLER_FOR_GET_REQUEST;
+import static org.apache.coyote.http11.handler.get.LoginPageHandler.LOGIN_PAGE_HANDLER;
 
 import java.util.Arrays;
 import org.apache.coyote.http11.HttpMethod;
